requirements.txt
Django==3.1.6
djangorestframework==3.12.2
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
redis==3.5.3
celery==4.4.7
flower==0.9.7
